Silence yet another warning
I have no clue why neither gcc nor clang throw a -Wsign-compare for me
there, but do for some other people, so... *sigh*

+++ src/common/compressor.cpp
@@ -243,7 +243,7 @@ void Compressor::writeData()
             return;
         }
 
-        if (_deflater->avail_out == ioBufferSize)
+        if (_deflater->avail_out == static_cast<unsigned int>(ioBufferSize))
             continue; // nothing to write here
 
         if (!_socket->write(_outputBuffer.constData(), ioBufferSize - _deflater->avail_out)) {
